Split visibility scoping parts of policies into their own classes
