+++ src/Access/DiscussionPolicy.php
@@ -11,38 +11,22 @@ namespace Flarum\Tags\Access;
 
 use Carbon\Carbon;
 use Flarum\Discussion\Discussion;
-use Flarum\Event\ScopeModelVisibility;
 use Flarum\Settings\SettingsRepositoryInterface;
-use Flarum\Tags\Tag;
-use Flarum\User\AbstractPolicy;
+use Flarum\User\Access\AbstractPolicy;
 use Flarum\User\User;
-use Illuminate\Contracts\Events\Dispatcher;
-use Illuminate\Database\Eloquent\Builder;
 
 class DiscussionPolicy extends AbstractPolicy
 {
-    /**
-     * {@inheritdoc}
-     */
-    protected $model = Discussion::class;
-
-    /**
-     * @var Dispatcher
-     */
-    protected $events;
-
     /**
      * @var SettingsRepositoryInterface
      */
     protected $settings;
 
     /**
-     * @param Dispatcher $events
      * @param SettingsRepositoryInterface $settings
      */
-    public function __construct(Dispatcher $events, SettingsRepositoryInterface $settings)
+    public function __construct(SettingsRepositoryInterface $settings)
     {
-        $this->events = $events;
         $this->settings = $settings;
     }
 
@@ -76,51 +60,6 @@ class DiscussionPolicy extends AbstractPolicy
                 return true;
             }
         }
-    }
-
-    /**
-     * @param User $actor
-     * @param Builder $query
-     */
-    public function find(User $actor, Builder $query)
-    {
-        // Hide discussions which have tags that the user is not allowed to see, unless an extension overrides this.
-        $query->where(function ($query) use ($actor) {
-            $query
-                ->whereNotIn('discussions.id', function ($query) use ($actor) {
-                    return $query->select('discussion_id')
-                        ->from('discussion_tag')
-                        ->whereIn('tag_id', Tag::getIdsWhereCannot($actor, 'viewDiscussions'));
-                })
-                ->orWhere(function ($query) use ($actor) {
-                    $this->events->dispatch(
-                        new ScopeModelVisibility($query, $actor, 'viewDiscussionsInRestrictedTags')
-                    );
-                });
-        });
-
-        // Hide discussions with no tags if the user doesn't have that global
-        // permission.
-        if (! $actor->hasPermission('viewDiscussions')) {
-            $query->has('tags');
-        }
-    }
-
-    /**
-     * @param User $actor
-     * @param Builder $query
-     * @param string $ability
-     */
-    public function findWithPermission(User $actor, Builder $query, $ability)
-    {
-        // If a discussion requires a certain permission in order for it to be
-        // visible, then we can check if the user has been granted that
-        // permission for any of the discussion's tags.
-        $query->whereIn('discussions.id', function ($query) use ($actor, $ability) {
-            return $query->select('discussion_id')
-                ->from('discussion_tag')
-                ->whereIn('tag_id', Tag::getIdsWhereCan($actor, 'discussion.'.$ability));
-        });
     }
 
     /**
