+++ src/Access/TagScopePolicy.php
@@ -1,0 +1,32 @@
+<?php
+
+/*
+ * This file is part of Flarum.
+ *
+ * For detailed copyright and license information, please view the
+ * LICENSE file that was distributed with this source code.
+ */
+
+namespace Flarum\Tags\Access;
+
+use Flarum\Tags\Tag;
+use Flarum\User\AbstractPolicy;
+use Flarum\User\User;
+use Illuminate\Database\Eloquent\Builder;
+
+class TagScopePolicy extends AbstractPolicy
+{
+    /**
+     * {@inheritdoc}
+     */
+    protected $model = Tag::class;
+
+    /**
+     * @param User $actor
+     * @param Builder $query
+     */
+    public function find(User $actor, Builder $query)
+    {
+        $query->whereNotIn('id', Tag::getIdsWhereCannot($actor, 'viewDiscussions'));
+    }
+}
